Put back git and openssh-client
Some CircleCI builds failed when building an annotated tag. See for
instance
https://circleci.com/gh/mozilla-services/kinto-dist/647. Builds seem
to succeed on non-annotated lightweight tags, but annotated tags are
more correct for doing releases. I believe this is because of bugs in
the CircleCI built-in git/ssh support and hope that restoring these
packages will cause the problem to go away.