Use machine for volume mounts (#244)
The idea is to switch from `docker` to `machine` in CircleCI. That gives us the ability to mount volumes. That means we can poke around for `mail/*.eml` (on the file system, even in CircleCI) instead of having to mess with a Python web server to do those queries. 

Being able to use volume mounts means that we now only need *1* `docker-compose.yml` file and it works pretty much identical in CircleCI as it does on your laptop.