Restore ec2-metadata dependency
This got dropped in #214, when listing dependencies with their
"extras" in setup.py was replaced with requirements.txt files. This is
one of two dependencies that were formerly pulled in by including
`kinto-fxa[scripts]` -- the other was boto3, which is already in
requirements/constraints.txt.

ec2-metadata itself has a missing dependency, so add it too.