Move testing Dockerfile to tests dir
We used to have the Dockerfiles for the server and tests in the same
directory. This is because we built the tests container on top of the
server, and our build setup required the containers to share the same
context.

Now, we build the containers with independent contexts. As a result, we
can move the testing Dockerfile to the tests directory. The only
additional changes needed to complete this migration were to:
- split out a testing-specific requirements.txt and include it in that
  directory
- add a named volume for database data so that in wouldn't be
  recreated during the tests after running the migrations