Support multiple dockerignores
To reuse the script to update + install system packages in both the
regular and integration test containers, we had to use the root
directory as the build context. This lead to a problem that we ignored
the `tests` directory for the regular container, but that's needed for
the integration test container. To get around this, we renamed both
Dockerfiles so that we could use multiple dockerignore files.