Remote Rust toolchains, make other Docker / docker compose changes (#362)
* Add canonicaljson-rs updated wheels to requirements.txt

* Tweak Dockerfiles

- remove rust toolchains now that we publish wheels that are compatible
   with most platforms
- Use virtual environment for `tests` Dockerfile
- remove pinned image hashes, since they are tied to specific platforms
- Add script for adding + updating system packages

Adapted from [here](https://pythonspeed.com/articles/system-packages-docker/)

To reuse the script to update + install system packages in both the
regular and integration test containers, we had to use the root
directory as the build context. This lead to an issue because we ignored
the `tests` directory for the regular container, but that's needed for
the integration test container. To get around this, we renamed both
Dockerfiles so that we could use multiple dockerignore files.

* Specify PGUSER in docker-compose

Prevents an error that says there's no role `root`