Reorder float error message so path is at the end, fixes #410
The `path` where a float is detected can be quite long.  Some tools like
Sentry can truncate the error message so the useful error message is
removed.

This makes troubleshooting more challenging and less obvious.  Moving the path to
the end makes it clearer what the error is and where to start looking
for the problem.